Refactor Client class to Application
